# [core] Speed up `EnsembleSelection._fit`