*Issue #, if available:*
The current implementation of the greedy ensembling is unnecessarily slowed down, especially for large ensemble_sizes, as the `weighted_ensemble_prediction` computation here
https://github.com/autogluon/autogluon/blob/f5c7b7400596e1449fad2b4107fde55809a09bac/core/src/autogluon/core/models/greedy_ensemble/ensemble_selection.py#L113-L125
is done in every iteration of the loop from scratch. This could be done on a rolling basis instead, which would reduce the complexity of this function from quadratic in the ensemble size to linear.

*Description of changes:*
This PR changes the current implementation to instead accumulate into the `ensemble_prediction` at each iteration. On a small local example I see a decrease in the ensembling time from 1.4s to 0.9s for the standard `ensemble_size=100`, and for `ensemble_size=1000` I see a decrease from 66s to 8.4s.


By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.