@LennartPurucker 2024-08-14
Great work! 

Two minor comments:

- I wonder how far the new code will be less numerically stable than the old code due to `ensemble_prediction *= (s - 1) / s`, any thoughts? 
- Also, if we need further speedups for ensemble selection, we can always consider parallelizing the internal for loop for roughly a `len(predictions) // n_jobs` speedup at the cost of much higher memory usage.