@Innixma 2024-08-14
@LennartPurucker I tested the PR on TabRepo, no noticeable numerical precision loss:

```
method & normalized-error & rank & time fit (s) & time infer (s) \\
PortfolioPR4367--N200 (ensemble) (4h) & 0.355 & 195.5 & 6335.7 & 0.036 \\
PortfolioMainline--N200 (ensemble) (4h) & 0.355 & 195.5 & 6335.7 & 0.036 \\
```